refactor: 전역 패키지 구조 정리 및 global 패키지 생성
- config, utils, exception 패키지를 global 하위로 이동
- 전역 설정 및 공통 로직을 global 계층에서 통합 관리하도록 구조 개선
- 패키지 import 경로 수정
