+++ src/main/java/com/sprint/mission/discodeit/global/config/RepositoryConfig.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.config;
+package com.sprint.mission.discodeit.global.config;
 
 import com.sprint.mission.discodeit.repository.*;
 import com.sprint.mission.discodeit.repository.file.*;
