+++ src/main/java/com/sprint/mission/discodeit/global/utils/TimeConvert.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.utils;
+package com.sprint.mission.discodeit.global.utils;
 
 import java.time.Instant;
 import java.time.LocalDateTime;
